+++ packages/queued-request-controller/src/QueuedRequestController.test.ts
@@ -25,6 +25,7 @@ describe('QueuedRequestController', () => {
   it('can be instantiated with default values', () => {
     const options: QueuedRequestControllerOptions = {
       messenger: buildQueuedRequestControllerMessenger(),
+      methodsRequiringNetworkSwitch: [],
     };
 
     const controller = new QueuedRequestController(options);
@@ -33,10 +34,7 @@ describe('QueuedRequestController', () => {
 
   describe('enqueueRequest', () => {
     it('skips the queue if the queue is empty and no request is being processed', async () => {
-      const options: QueuedRequestControllerOptions = {
-        messenger: buildQueuedRequestControllerMessenger(),
-      };
-      const controller = new QueuedRequestController(options);
+      const controller = buildQueuedRequestController();
 
       await controller.enqueueRequest(buildRequest(), async () => {
         expect(controller.state.queuedRequestCount).toBe(0);
@@ -45,10 +43,7 @@ describe('QueuedRequestController', () => {
     });
 
     it('skips the queue if the queue is empty and the request being processed has the same origin', async () => {
-      const options: QueuedRequestControllerOptions = {
-        messenger: buildQueuedRequestControllerMessenger(),
-      };
-      const controller = new QueuedRequestController(options);
+      const controller = buildQueuedRequestController();
       // Trigger first request
       const firstRequest = controller.enqueueRequest(
         buildRequest(),
@@ -65,7 +60,7 @@ describe('QueuedRequestController', () => {
       await firstRequest;
     });
 
-    it('switches network if a request comes in for a different network client', async () => {
+    it('switches network if a request comes in for a different network client and the method is in the methodsRequiringNetworkSwitch param', async () => {
       const mockSetActiveNetwork = jest.fn();
       const { messenger } = buildControllerMessenger({
         networkControllerGetState: jest.fn().mockReturnValue({
@@ -82,13 +77,13 @@ describe('QueuedRequestController', () => {
         'QueuedRequestController:networkSwitched',
         onNetworkSwitched,
       );
-      const options: QueuedRequestControllerOptions = {
+      const controller = buildQueuedRequestController({
         messenger: buildQueuedRequestControllerMessenger(messenger),
-      };
-      const controller = new QueuedRequestController(options);
+        methodsRequiringNetworkSwitch: ['method_requiring_network_switch'],
+      });
 
       await controller.enqueueRequest(
-        buildRequest(),
+        { ...buildRequest(), method: 'method_requiring_network_switch' },
         () => new Promise((resolve) => setTimeout(resolve, 10)),
       );
 
@@ -100,7 +95,7 @@ describe('QueuedRequestController', () => {
       );
     });
 
-    it('does not switch networks if the method is `eth_requestAccounts`', async () => {
+    it('does not switch networks if the method is not in the methodsRequiringNetworkSwitch param', async () => {
       const mockSetActiveNetwork = jest.fn();
       const { messenger } = buildControllerMessenger({
         networkControllerGetState: jest.fn().mockReturnValue({
@@ -117,13 +112,13 @@ describe('QueuedRequestController', () => {
         'QueuedRequestController:networkSwitched',
         onNetworkSwitched,
       );
-      const options: QueuedRequestControllerOptions = {
+      const controller = buildQueuedRequestController({
         messenger: buildQueuedRequestControllerMessenger(messenger),
-      };
-      const controller = new QueuedRequestController(options);
+        methodsRequiringNetworkSwitch: [],
+      });
 
       await controller.enqueueRequest(
-        { ...buildRequest(), method: 'eth_requestAccounts' },
+        { ...buildRequest(), method: 'not_in_methodsRequiringNetworkSwitch' },
         () => new Promise((resolve) => setTimeout(resolve, 10)),
       );
 
@@ -148,10 +143,9 @@ describe('QueuedRequestController', () => {
         'QueuedRequestController:networkSwitched',
         onNetworkSwitched,
       );
-      const options: QueuedRequestControllerOptions = {
+      const controller = buildQueuedRequestController({
         messenger: buildQueuedRequestControllerMessenger(messenger),
-      };
-      const controller = new QueuedRequestController(options);
+      });
 
       await controller.enqueueRequest(
         buildRequest(),
@@ -163,10 +157,7 @@ describe('QueuedRequestController', () => {
     });
 
     it('queues request if a request from another origin is being processed', async () => {
-      const options: QueuedRequestControllerOptions = {
-        messenger: buildQueuedRequestControllerMessenger(),
-      };
-      const controller = new QueuedRequestController(options);
+      const controller = buildQueuedRequestController();
       // Trigger first request
       const firstRequest = controller.enqueueRequest(
         { ...buildRequest(), origin: 'https://exampleorigin1.metamask.io' },
@@ -189,10 +180,7 @@ describe('QueuedRequestController', () => {
     });
 
     it('drains batch from queue when current batch finishes', async () => {
-      const options: QueuedRequestControllerOptions = {
-        messenger: buildQueuedRequestControllerMessenger(),
-      };
-      const controller = new QueuedRequestController(options);
+      const controller = buildQueuedRequestController();
       // Trigger first batch
       const firstRequest = controller.enqueueRequest(
         { ...buildRequest(), origin: 'https://firstbatch.metamask.io' },
@@ -236,10 +224,7 @@ describe('QueuedRequestController', () => {
     });
 
     it('drains batch from queue when current batch finishes with requests out-of-order', async () => {
-      const options: QueuedRequestControllerOptions = {
-        messenger: buildQueuedRequestControllerMessenger(),
-      };
-      const controller = new QueuedRequestController(options);
+      const controller = buildQueuedRequestController();
       // Trigger first batch
       const firstRequest = controller.enqueueRequest(
         { ...buildRequest(), origin: 'https://firstbatch.metamask.io' },
@@ -283,10 +268,7 @@ describe('QueuedRequestController', () => {
     });
 
     it('processes requests from each batch in parallel', async () => {
-      const options: QueuedRequestControllerOptions = {
-        messenger: buildQueuedRequestControllerMessenger(),
-      };
-      const controller = new QueuedRequestController(options);
+      const controller = buildQueuedRequestController();
       const firstRequest = controller.enqueueRequest(
         { ...buildRequest(), origin: 'https://firstorigin.metamask.io' },
         async () => {
@@ -342,10 +324,7 @@ describe('QueuedRequestController', () => {
     });
 
     it('preserves request order within each batch', async () => {
-      const options: QueuedRequestControllerOptions = {
-        messenger: buildQueuedRequestControllerMessenger(),
-      };
-      const controller = new QueuedRequestController(options);
+      const controller = buildQueuedRequestController();
       const executionOrder: string[] = [];
       const firstRequest = controller.enqueueRequest(
         { ...buildRequest(), origin: 'https://firstorigin.metamask.io' },
@@ -398,10 +377,7 @@ describe('QueuedRequestController', () => {
     });
 
     it('preserves request order even when interlaced with requests from other origins', async () => {
-      const options: QueuedRequestControllerOptions = {
-        messenger: buildQueuedRequestControllerMessenger(),
-      };
-      const controller = new QueuedRequestController(options);
+      const controller = buildQueuedRequestController();
       const executionOrder: string[] = [];
       const firstRequest = controller.enqueueRequest(
         { ...buildRequest(), origin: 'https://firstorigin.metamask.io' },
@@ -461,10 +437,9 @@ describe('QueuedRequestController', () => {
         'QueuedRequestController:networkSwitched',
         onNetworkSwitched,
       );
-      const options: QueuedRequestControllerOptions = {
+      const controller = buildQueuedRequestController({
         messenger: buildQueuedRequestControllerMessenger(messenger),
-      };
-      const controller = new QueuedRequestController(options);
+      });
       const firstRequest = controller.enqueueRequest(
         { ...buildRequest(), origin: 'https://firstorigin.metamask.io' },
         () => new Promise((resolve) => setTimeout(resolve, 10)),
@@ -513,10 +488,9 @@ describe('QueuedRequestController', () => {
         'QueuedRequestController:networkSwitched',
         onNetworkSwitched,
       );
-      const options: QueuedRequestControllerOptions = {
+      const controller = buildQueuedRequestController({
         messenger: buildQueuedRequestControllerMessenger(messenger),
-      };
-      const controller = new QueuedRequestController(options);
+      });
       const firstRequest = controller.enqueueRequest(
         { ...buildRequest(), origin: 'firstorigin.metamask.io' },
         () => new Promise((resolve) => setTimeout(resolve, 10)),
@@ -558,14 +532,18 @@ describe('QueuedRequestController', () => {
             .fn()
             .mockImplementation((_origin) => 'differentNetworkClientId'),
         });
-        const options: QueuedRequestControllerOptions = {
+        const controller = buildQueuedRequestController({
           messenger: buildQueuedRequestControllerMessenger(messenger),
-        };
-        const controller = new QueuedRequestController(options);
+          methodsRequiringNetworkSwitch: ['method_requiring_network_switch'],
+        });
 
         await expect(() =>
           controller.enqueueRequest(
-            { ...buildRequest(), origin: 'https://example.metamask.io' },
+            {
+              ...buildRequest(),
+              method: 'method_requiring_network_switch',
+              origin: 'https://example.metamask.io',
+            },
             jest.fn(),
           ),
         ).rejects.toThrow(switchError);
@@ -589,12 +567,16 @@ describe('QueuedRequestController', () => {
                 : 'selectedNetworkClientId',
             ),
         });
-        const options: QueuedRequestControllerOptions = {
+        const controller = buildQueuedRequestController({
           messenger: buildQueuedRequestControllerMessenger(messenger),
-        };
-        const controller = new QueuedRequestController(options);
+          methodsRequiringNetworkSwitch: ['method_requiring_network_switch'],
+        });
         const firstRequest = controller.enqueueRequest(
-          { ...buildRequest(), origin: 'https://firstorigin.metamask.io' },
+          {
+            ...buildRequest(),
+            method: 'method_requiring_network_switch',
+            origin: 'https://firstorigin.metamask.io',
+          },
           () => new Promise((resolve) => setTimeout(resolve, 10)),
         );
         // ensure first request skips queue
@@ -605,7 +587,11 @@ describe('QueuedRequestController', () => {
             () => new Promise((resolve) => setTimeout(resolve, 100)),
           );
         const secondRequest = controller.enqueueRequest(
-          { ...buildRequest(), origin: 'https://secondorigin.metamask.io' },
+          {
+            ...buildRequest(),
+            method: 'method_requiring_network_switch',
+            origin: 'https://secondorigin.metamask.io',
+          },
           secondRequestNext,
         );
 
@@ -635,12 +621,16 @@ describe('QueuedRequestController', () => {
                 : 'selectedNetworkClientId',
             ),
         });
-        const options: QueuedRequestControllerOptions = {
+        const controller = buildQueuedRequestController({
           messenger: buildQueuedRequestControllerMessenger(messenger),
-        };
-        const controller = new QueuedRequestController(options);
+          methodsRequiringNetworkSwitch: ['method_requiring_network_switch'],
+        });
         const firstRequest = controller.enqueueRequest(
-          { ...buildRequest(), origin: 'https://firstorigin.metamask.io' },
+          {
+            ...buildRequest(),
+            method: 'method_requiring_network_switch',
+            origin: 'https://firstorigin.metamask.io',
+          },
           () => new Promise((resolve) => setTimeout(resolve, 10)),
         );
         // ensure first request skips queue
@@ -651,7 +641,11 @@ describe('QueuedRequestController', () => {
             () => new Promise((resolve) => setTimeout(resolve, 100)),
           );
         const secondRequest = controller.enqueueRequest(
-          { ...buildRequest(), origin: 'https://secondorigin.metamask.io' },
+          {
+            ...buildRequest(),
+            method: 'method_requiring_network_switch',
+            origin: 'https://secondorigin.metamask.io',
+          },
           secondRequestNext,
         );
         // ensure test starts with one request queued up
@@ -680,12 +674,16 @@ describe('QueuedRequestController', () => {
                 : 'selectedNetworkClientId',
             ),
         });
-        const options: QueuedRequestControllerOptions = {
+        const controller = buildQueuedRequestController({
           messenger: buildQueuedRequestControllerMessenger(messenger),
-        };
-        const controller = new QueuedRequestController(options);
+          methodsRequiringNetworkSwitch: ['method_requiring_network_switch'],
+        });
         const firstRequest = controller.enqueueRequest(
-          { ...buildRequest(), origin: 'https://firstorigin.metamask.io' },
+          {
+            ...buildRequest(),
+            method: 'method_requiring_network_switch',
+            origin: 'https://firstorigin.metamask.io',
+          },
           () => new Promise((resolve) => setTimeout(resolve, 10)),
         );
         // ensure first request skips queue
@@ -696,7 +694,11 @@ describe('QueuedRequestController', () => {
             () => new Promise((resolve) => setTimeout(resolve, 100)),
           );
         const secondRequest = controller.enqueueRequest(
-          { ...buildRequest(), origin: 'https://secondorigin.metamask.io' },
+          {
+            ...buildRequest(),
+            method: 'method_requiring_network_switch',
+            origin: 'https://secondorigin.metamask.io',
+          },
           secondRequestNext,
         );
         const thirdRequestNext = jest
@@ -705,7 +707,11 @@ describe('QueuedRequestController', () => {
             () => new Promise((resolve) => setTimeout(resolve, 100)),
           );
         const thirdRequest = controller.enqueueRequest(
-          { ...buildRequest(), origin: 'https://thirdorigin.metamask.io' },
+          {
+            ...buildRequest(),
+            method: 'method_requiring_network_switch',
+            origin: 'https://thirdorigin.metamask.io',
+          },
           thirdRequestNext,
         );
         // ensure test starts with two requests queued up
@@ -722,11 +728,7 @@ describe('QueuedRequestController', () => {
 
     describe('when a request fails', () => {
       it('throws error', async () => {
-        const options: QueuedRequestControllerOptions = {
-          messenger: buildQueuedRequestControllerMessenger(),
-        };
-
-        const controller = new QueuedRequestController(options);
+        const controller = buildQueuedRequestController();
 
         // Mock a request that throws an error
         const requestWithError = jest.fn(() =>
@@ -744,10 +746,7 @@ describe('QueuedRequestController', () => {
       });
 
       it('correctly updates the request queue count upon failure', async () => {
-        const options: QueuedRequestControllerOptions = {
-          messenger: buildQueuedRequestControllerMessenger(),
-        };
-        const controller = new QueuedRequestController(options);
+        const controller = buildQueuedRequestController();
 
         await expect(() =>
           controller.enqueueRequest(
@@ -761,11 +760,7 @@ describe('QueuedRequestController', () => {
       });
 
       it('correctly processes the next item in the queue', async () => {
-        const options: QueuedRequestControllerOptions = {
-          messenger: buildQueuedRequestControllerMessenger(),
-        };
-
-        const controller = new QueuedRequestController(options);
+        const controller = buildQueuedRequestController();
 
         // Mock requests with one request throwing an error
         const request1 = jest.fn(async () => {
@@ -898,6 +893,24 @@ function buildQueuedRequestControllerMessenger(
     ],
     allowedEvents: [],
   });
+}
+
+/**
+ * Builds a QueuedRequestController
+ *
+ * @param overrideOptions - The optional options object.
+ * @returns The QueuedRequestController.
+ */
+function buildQueuedRequestController(
+  overrideOptions?: Partial<QueuedRequestControllerOptions>,
+): QueuedRequestController {
+  const options: QueuedRequestControllerOptions = {
+    messenger: buildQueuedRequestControllerMessenger(),
+    methodsRequiringNetworkSwitch: [],
+    ...overrideOptions,
+  };
+
+  return new QueuedRequestController(options);
 }
 
 /**